v1.0.1: has no rights all

+++ pkg/service/ecomorph-entity-service.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	context "context"
+	er "errors"
 	"github.com/infobloxopen/atlas-app-toolkit/v2/rpc/resource"
 	"github.com/sergey23144V/BotanyBackend/pkg"
 	"github.com/sergey23144V/BotanyBackend/pkg/middlewares"
@@ -28,7 +29,11 @@ func NewEcomorphsEntityServiceImpl(repository *repository.Repository) EcomorphsE
 }
 
 func (e EcomorphsEntityServiceImpl) CreateEcomorphsEntity(ctx context.Context, entity *api.InputEcomorphsEntity) (*api.EcomorphsEntity, error) {
-	return e.repository.CreateEcomorphsEntity(ctx, e.ToPB(ctx, entity))
+	pb, err := e.ToPB(ctx, entity)
+	if err != nil {
+		return nil, err
+	}
+	return e.repository.CreateEcomorphsEntity(ctx, pb)
 }
 
 func (e EcomorphsEntityServiceImpl) GetEcomorphsEntityById(ctx context.Context, request *api.IdRequest) (*api.EcomorphsEntity, error) {
@@ -39,7 +44,8 @@ func (e EcomorphsEntityServiceImpl) GetEcomorphsEntityById(ctx context.Context, 
 func (e EcomorphsEntityServiceImpl) DeleteEcomorphsEntity(ctx context.Context, request *api.IdRequest) (*api.BoolResponse, error) {
 	userId := middlewares.GetUserIdFromContext(ctx)
 	result := &api.BoolResponse{Result: true}
-	err := e.repository.EcomorphsEntityRepository.DeleteEcomorphsEntity(ctx, &api.EcomorphsEntity{Id: request.Id, UserId: userId})
+	role := middlewares.GetUserRoleFromContext(ctx)
+	err := e.repository.EcomorphsEntityRepository.DeleteEcomorphsEntity(ctx, &api.EcomorphsEntity{Id: request.Id, UserId: userId}, *role)
 	if err != nil {
 		result.Result = false
 		return result, err
@@ -48,7 +54,11 @@ func (e EcomorphsEntityServiceImpl) DeleteEcomorphsEntity(ctx context.Context, r
 }
 
 func (e EcomorphsEntityServiceImpl) StrictUpdateEcomorphsEntity(ctx context.Context, entity *api.InputEcomorphsEntity) (*api.EcomorphsEntity, error) {
-	return e.repository.EcomorphsEntityRepository.StrictUpdateEcomorphsEntity(ctx, e.ToPB(ctx, entity))
+	pb, err := e.ToPB(ctx, entity)
+	if err != nil {
+		return nil, err
+	}
+	return e.repository.EcomorphsEntityRepository.StrictUpdateEcomorphsEntity(ctx, pb)
 }
 
 func (e EcomorphsEntityServiceImpl) UpdateEcomorphsEntity(ctx context.Context, entity *api.InputEcomorphsEntity) (*api.EcomorphsEntity, error) {
@@ -59,7 +69,12 @@ func (e EcomorphsEntityServiceImpl) UpdateEcomorphsEntity(ctx context.Context, e
 	if entity.Input.Description != "" {
 		fieldMask = append(fieldMask, "Description")
 	}
-	return e.repository.EcomorphsEntityRepository.UpdateEcomorphsEntity(ctx, e.ToPB(ctx, entity), &field_mask.FieldMask{Paths: fieldMask})
+	pb, err := e.ToPB(ctx, entity)
+	if err != nil {
+		return nil, err
+	}
+	role := middlewares.GetUserRoleFromContext(ctx)
+	return e.repository.EcomorphsEntityRepository.UpdateEcomorphsEntity(ctx, pb, &field_mask.FieldMask{Paths: fieldMask}, *role)
 }
 
 func (e EcomorphsEntityServiceImpl) GetListEcomorphsEntity(ctx context.Context, request *api.EcomorphsEntityListRequest) (*api.EcomorphsEntityList, error) {
@@ -68,14 +83,17 @@ func (e EcomorphsEntityServiceImpl) GetListEcomorphsEntity(ctx context.Context, 
 	if err != nil {
 		return nil, err
 	}
-	if request != nil {
+	if request.Page != nil {
 		page = &api.PagesResponse{Page: request.Page.Page, Limit: request.Page.Limit, Total: int32(len(list))}
 	}
 	return &api.EcomorphsEntityList{List: list, Page: page}, nil
 }
 
-func (e EcomorphsEntityServiceImpl) ToPB(ctx context.Context, entity *api.InputEcomorphsEntity) *api.EcomorphsEntity {
-	var id *resource.Identifier
+func (e EcomorphsEntityServiceImpl) ToPB(ctx context.Context, entity *api.InputEcomorphsEntity) (*api.EcomorphsEntity, error) {
+	var (
+		id             *resource.Identifier
+		ecomorphEntity *api.EcomorphsEntity
+	)
 
 	if entity.Id != nil {
 		id = entity.Id
@@ -83,13 +101,28 @@ func (e EcomorphsEntityServiceImpl) ToPB(ctx context.Context, entity *api.InputE
 		id = &resource.Identifier{ResourceId: pkg.GenerateUUID()}
 	}
 	userId := middlewares.GetUserIdFromContext(ctx)
-	return &api.EcomorphsEntity{
-		Id:           id,
-		Title:        entity.Input.Title,
-		Description:  entity.Input.Description,
-		Ecomorphs:    entity.Input.Ecomorphs,
-		Score:        entity.Input.Score,
-		DisplayTable: entity.Input.DisplayTable,
-		UserId:       userId,
+	role := middlewares.GetUserRoleFromContext(ctx)
+	if *role == api.RoleType_SuperUser && entity.Publicly {
+		ecomorphEntity = &api.EcomorphsEntity{
+			Id:           id,
+			Title:        entity.Input.Title,
+			Description:  entity.Input.Description,
+			Ecomorphs:    entity.Input.Ecomorphs,
+			Score:        entity.Input.Score,
+			DisplayTable: entity.Input.DisplayTable,
+		}
+	} else if *role != api.RoleType_SuperUser && entity.Publicly {
+		return nil, er.New("has no rights")
+	} else {
+		ecomorphEntity = &api.EcomorphsEntity{
+			Id:           id,
+			Title:        entity.Input.Title,
+			Description:  entity.Input.Description,
+			Ecomorphs:    entity.Input.Ecomorphs,
+			Score:        entity.Input.Score,
+			DisplayTable: entity.Input.DisplayTable,
+			UserId:       userId,
+		}
 	}
+	return ecomorphEntity, nil
 }
