+++ pkg/service/type-plant-service.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	er "errors"
 	"github.com/infobloxopen/atlas-app-toolkit/v2/rpc/resource"
 	"github.com/sergey23144V/BotanyBackend/servers/g-rpc/api"
 	"google.golang.org/genproto/protobuf/field_mask"
@@ -30,7 +31,11 @@ func NewTypePlantServiceImpl(repository *repository.Repository) TypePlantService
 }
 
 func (t TypePlantServiceImpl) CreateTypePlant(ctx context.Context, ecomorph *api.InputTypePlantRequest) (*api.TypePlant, error) {
-	return t.repository.TypePlantRepository.CreateTypePlant(ctx, t.ToPB(ctx, ecomorph))
+	pb, err := t.ToPB(ctx, ecomorph)
+	if err != nil {
+		return nil, err
+	}
+	return t.repository.TypePlantRepository.CreateTypePlant(ctx, pb)
 }
 
 func (t TypePlantServiceImpl) GetTypePlantById(ctx context.Context, request *api.IdRequest) (*api.TypePlant, error) {
@@ -41,7 +46,8 @@ func (t TypePlantServiceImpl) GetTypePlantById(ctx context.Context, request *api
 func (t TypePlantServiceImpl) DeleteTypePlant(ctx context.Context, request *api.IdRequest) (*api.BoolResponse, error) {
 	userId := middlewares.GetUserIdFromContext(ctx)
 	result := &api.BoolResponse{Result: true}
-	err := t.repository.TypePlantRepository.DeleteTypePlant(ctx, &api.TypePlant{Id: request.Id, UserId: userId})
+	role := middlewares.GetUserRoleFromContext(ctx)
+	err := t.repository.TypePlantRepository.DeleteTypePlant(ctx, &api.TypePlant{Id: request.Id, UserId: userId}, *role)
 	if err != nil {
 		result.Result = false
 		return nil, err
@@ -50,7 +56,11 @@ func (t TypePlantServiceImpl) DeleteTypePlant(ctx context.Context, request *api.
 }
 
 func (t TypePlantServiceImpl) StrictUpdateTypePlant(ctx context.Context, in *api.InputTypePlantRequest) (*api.TypePlant, error) {
-	return t.repository.TypePlantRepository.StrictUpdateTypePlant(ctx, t.ToPB(ctx, in))
+	pb, err := t.ToPB(ctx, in)
+	if err != nil {
+		return nil, err
+	}
+	return t.repository.TypePlantRepository.StrictUpdateTypePlant(ctx, pb)
 }
 
 func (t TypePlantServiceImpl) UpdateTypePlant(ctx context.Context, in *api.InputTypePlantRequest) (*api.TypePlant, error) {
@@ -61,19 +71,25 @@ func (t TypePlantServiceImpl) UpdateTypePlant(ctx context.Context, in *api.Input
 	if in.Input.Subtitle != "" {
 		fieldMask = append(fieldMask, "Subtitle")
 	}
-	return t.repository.TypePlantRepository.UpdateTypePlant(ctx, t.ToPB(ctx, in), &field_mask.FieldMask{Paths: fieldMask})
+	role := middlewares.GetUserRoleFromContext(ctx)
+	pb, err := t.ToPB(ctx, in)
+	if err != nil {
+		return nil, err
+	}
+	return t.repository.TypePlantRepository.UpdateTypePlant(ctx, pb, &field_mask.FieldMask{Paths: fieldMask}, *role)
 }
 
 func (t TypePlantServiceImpl) AddEcomorphsEntityToTypePlant(ctx context.Context, in *api.InputTypePlant_EcomorphsEntityRequest) (*api.TypePlant, error) {
 	userId := middlewares.GetUserIdFromContext(ctx)
-	return t.repository.TypePlantRepository.UpdateTypePlant(ctx, &api.TypePlant{Id: in.Id, EcomorphsEntity: in.EcomorphsEntity, UserId: userId}, &field_mask.FieldMask{Paths: []string{}})
+	role := middlewares.GetUserRoleFromContext(ctx)
+	return t.repository.TypePlantRepository.UpdateTypePlant(ctx, &api.TypePlant{Id: in.Id, EcomorphsEntity: in.EcomorphsEntity, UserId: userId}, &field_mask.FieldMask{Paths: []string{}}, *role)
 }
 
 func (t TypePlantServiceImpl) GetListTypePlant(ctx context.Context, request *api.TypePlantListRequest) (*api.TypePlantList, error) {
 	var page *api.PagesResponse
 	userId := middlewares.GetUserIdFromContext(ctx)
 	getList, err := t.repository.TypePlantRepository.GetListTypePlant(ctx, &api.TypePlant{UserId: userId}, request)
-	if request != nil {
+	if request.Page != nil {
 		page = &api.PagesResponse{Page: request.Page.Page, Limit: request.Page.Limit, Total: int32(len(getList))}
 	}
 	if err != nil {
@@ -83,7 +99,7 @@ func (t TypePlantServiceImpl) GetListTypePlant(ctx context.Context, request *api
 	return result, nil
 }
 
-func (t TypePlantServiceImpl) ToPB(ctx context.Context, request *api.InputTypePlantRequest) *api.TypePlant {
+func (t TypePlantServiceImpl) ToPB(ctx context.Context, request *api.InputTypePlantRequest) (*api.TypePlant, error) {
 	var id *resource.Identifier
 
 	if request.Id != nil {
@@ -93,12 +109,26 @@ func (t TypePlantServiceImpl) ToPB(ctx context.Context, request *api.InputTypePl
 	}
 
 	userId := middlewares.GetUserIdFromContext(ctx)
-	return &api.TypePlant{
-		Id:              id,
-		Title:           request.Input.Title,
-		Subtitle:        request.Input.Subtitle,
-		EcomorphsEntity: request.Input.EcomorphsEntity,
-		Img:             request.Input.Img,
-		UserId:          userId,
+	role := middlewares.GetUserRoleFromContext(ctx)
+	if *role == api.RoleType_SuperUser && request.Publicly {
+		return &api.TypePlant{
+			Id:              id,
+			Title:           request.Input.Title,
+			Subtitle:        request.Input.Subtitle,
+			EcomorphsEntity: request.Input.EcomorphsEntity,
+			Img:             request.Input.Img,
+		}, nil
+	} else if *role != api.RoleType_SuperUser && request.Publicly {
+		return nil, er.New("has no rights")
+	} else {
+		return &api.TypePlant{
+			Id:              id,
+			Title:           request.Input.Title,
+			Subtitle:        request.Input.Subtitle,
+			EcomorphsEntity: request.Input.EcomorphsEntity,
+			Img:             request.Input.Img,
+			UserId:          userId,
+		}, nil
 	}
+
 }
